requirements.txt
chromadb
langchain
langchain-community
google-generativeai
sentence-transformers
python-dotenv
tqdm
numpy
